% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBox.R
\name{checkBox}
\alias{checkBox}
\title{Check if flasks in a box have been recorded as sampled from the autosampler}
\usage{
checkBox(samplerBoxData, controlFile)
}
\arguments{
\item{samplerBoxData}{A data.frame of sampled boxes produced as a report from
the autosampler and loaded using the function selectBox}

\item{controlFile}{a list with control file data produced by load_cf}
}
\value{
a vector with the position of the flasks that match the sampler box data.
}
\description{
Check if flasks in a box have been recorded as sampled from the autosampler
}
