% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{flag}
\alias{flag}
\title{Flags for time series data}
\usage{
flag(ts, outliers)
}
\arguments{
\item{ts}{a numeric time series for a gas species}

\item{outliers}{a vector of outliers identified by the function \code{\link[=smoothingSpline]{smoothingSpline()}}}
}
\value{
an integer vector with values of 0 for observations that pass the
time series test, and a numeric value of 1 for observations that do not pass
the test. Observations with a flag of 1 are not representative of a regional
scale background
}
\description{
Flags for data based on a time series test. The input to the function should
be outliers identified by the fitting of a smoothing spline with the function
\code{\link[=smoothingSpline]{smoothingSpline()}}. A flag of 0 indicates good data, and a flag value of 1
indicates failure of the time series test.
}
