% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBox.R
\name{getBox}
\alias{getBox}
\title{Get box data and metadata from control file}
\usage{
getBox(cf, box)
}
\arguments{
\item{cf}{control file as list. This is produced by function load_cf}

\item{box}{character string with the ID of the box to extract}
}
\value{
a list with data and metadata for a single box
}
\description{
Get box data and metadata from control file
}
