% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampledBox.R
\name{getSampledBox}
\alias{getSampledBox}
\title{Get information of a sampled box from a report from the autosampler}
\usage{
getSampledBox(
  file,
  box,
  selvars = c("sampleID", "flaskID", "batchID", "boxID", "startTime", "endTime",
    "flaskPort", "flaskInsertTime", "mountPressure", "samplePressure")
)
}
\arguments{
\item{file}{a character string with the location of the csv file containing information from the autosampler. 
Usually a result to a call to the function sampledBoxes}

\item{box}{character string with the Box ID}

\item{selvars}{character vector of variables to select from original file}
}
\value{
a data.frame with information on flasks sampled for the particular box
}
\description{
Get information of a sampled box from a report from the autosampler
}
\examples{
\dontrun{
getSampledBox(file="FlaskCycle/packedBoxes/sampledBoxes.csv", box="B0171")
}
}
