% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGas.R
\name{plotGas}
\alias{plotGas}
\title{Plot data for a specific gas species}
\usage{
plotGas(
  x,
  gas,
  start = "2021-06-01",
  end = "2023-12-31",
  sigma = NULL,
  type = "b",
  flag = NULL,
  ...
)
}
\arguments{
\item{x}{release data.frame}

\item{gas}{character string with the name of the gas to plot}

\item{start}{character string with the start date for the horizonta axis}

\item{end}{character string with the end date for the horizontal axis}

\item{sigma}{character string with the name of the variable that represents the uncertainty in the data}

\item{type}{argument passed to plot() for the type of plot}

\item{flag}{vector of integers with observations to flag}

\item{...}{other arguments passed to plot()}
}
\value{
a plot
}
\description{
Plot data for a specific gas species
}
