% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIsoLab.R
\name{readIsoLab}
\alias{readIsoLab}
\title{Read file delivered by the IsoLab and select specific columns with new names}
\usage{
readIsoLab(
  file,
  columns,
  vars = c("USN", "request number", "FlaskNo", "box", "sampling time", "O2_N2",
    "sigma_O2_N2", "Ar_N2", "sigma_Ar_N2", "d13C-CO2", "d18O-CO2", "sigma_d13C-CO2",
    "sigma_d18O-CO2", "d13C-CH4", "d2H-CH4", "sigma_d13C-CH4", "sigma_d2H-CH4")
)
}
\arguments{
\item{file}{character string with the path to the location of the file to read.
Extension must be in .xls or .xlsx}

\item{columns}{character vector with the column letters that correspond to the variables to select.}

\item{vars}{character vector with the new names to assign to selected variables.}
}
\value{
a data.frame with the selected variables
}
\description{
Read file delivered by the IsoLab and select specific columns with new names
}
