% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledBoxes.R
\name{sampledBoxes}
\alias{sampledBoxes}
\title{Collect all information from the report produced by the autosampler about 
packed boxes and prints a report}
\usage{
sampledBoxes(
  fromWhere = "~/Repos/ATTO_flask_sampling/FlaskCycle/packedBoxes",
  toWhere = fromWhere
)
}
\arguments{
\item{fromWhere}{a character string with the path with the reports are stored.}

\item{toWhere}{a character string with the path to store the report.}
}
\value{
a csv file with the data of all sampled boxes
}
\description{
Collect all information from the report produced by the autosampler about 
packed boxes and prints a report
}
\examples{
sampledBoxes()
}
