% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothLines.R
\name{smoothLines}
\alias{smoothLines}
\title{Add smooth lines to plot}
\usage{
smoothLines(pred.spline, ...)
}
\arguments{
\item{pred.spline}{a data.frame of predictions obtained from \link{smoothingSpline}}

\item{...}{other ploting arguments pass to \code{\link[graphics:lines]{graphics::lines()}}.}
}
\value{
lines added to a plot
}
\description{
Takes predictions from a smoothing spline obtained with the function
\code{\link[=smoothingSpline]{smoothingSpline()}} and add smooth lines to a plot produced with the function
\code{\link[=plotGas]{plotGas()}}.
}
