% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothPolygon.R
\name{smoothPolygon}
\alias{smoothPolygon}
\title{Adds polygon around smoothing spline in plot}
\usage{
smoothPolygon(
  pred.spline,
  rsme,
  col = gray(0.5, alpha = 0.3),
  border = gray(0.5),
  ...
)
}
\arguments{
\item{pred.spline}{the output obtained from \link{smoothingSpline}}

\item{rsme}{a value of the rsme obtained from the output of a call to \code{\link[=smoothingSpline]{smoothingSpline()}}}

\item{...}{other ploting arguments pass to \code{\link[graphics:polygon]{graphics::polygon()}}.}
}
\value{
lines added to a plot
}
\description{
Takes predictions from a smoothing spline obtained with the function
\code{\link[=smoothingSpline]{smoothingSpline()}} plus the root mean square error of the fitting
and adds a polygon to a plot produced with the function
\code{\link[=plotGas]{plotGas()}}.
}
